backend/ev_charging/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiohttp-retry==2.9.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.9.0
astroid==3.3.11
attrs==25.3.0
av==15.0.0
cachetools==5.5.2
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
cryptography==45.0.5
dill==0.4.0
distro==1.9.0
Django==5.2.4
django-cors-headers==4.7.0
django-redis==6.0.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.2.2
docstring_parser==0.16
eval_type_backport==0.2.2
filetype==1.2.0
flatbuffers==25.2.10
frozenlist==1.7.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.181.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-speech==2.33.0
google-cloud-texttospeech==2.29.0
google-genai==1.36.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.74.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.30.2
httpx==0.28.1
humanfriendly==10.0
idna==3.10
isort==6.0.1
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jwt==1.4.0
langchain-core==0.3.76
langchain-google-genai==2.1.10
langchain-openai==0.3.33
langchain-pinecone==0.2.12
langsmith==0.4.25
livekit==1.0.9
livekit-agents==1.1.4
livekit-api==1.0.2
livekit-plugins-deepgram==1.1.4
livekit-plugins-elevenlabs==1.1.4
livekit-plugins-google==1.1.4
livekit-plugins-openai==1.1.4
livekit-plugins-silero==1.1.4
livekit-protocol==1.0.3
mccabe==0.7.0
mpmath==1.3.0
multidict==6.6.3
nest-asyncio==1.6.0
numpy==2.2.1
onnxruntime==1.22.0
openai==1.94.0
orjson==3.11.3
packaging==25.0
pillow==11.3.0
pinecone==7.3.0
pinecone-plugin-assistant==1.8.0
pinecone-plugin-interface==0.0.7
platformdirs==4.4.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.31.1
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.8.0
pylint==3.3.8
pylint-django==2.6.1
pylint-plugin-utils==0.9.0
pyparsing==3.2.3
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.2
redis==6.4.0
regex==2025.9.1
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
simsimd==6.5.3
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.2
sqlparse==0.5.3
sympy==1.14.0
tenacity==9.1.2
tiktoken==0.11.0
tomlkit==0.13.3
tqdm==4.67.1
types-protobuf==4.25.0.20240417
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
watchfiles==1.1.0
websockets==15.0.1
yarl==1.20.1
zstandard==0.24.0
